# public/swagger/swagger.yaml
swagger: '2.0'
info:
  description: Mathews Onyango API Laboratory- RECIPE API PROJECT
  version: 1.0.0
  title: Mathews Onyango API Laboratory - RECIPE API PROJECT
basePath: /api/
...
paths:
...
  /events/{event}:
    get:
      tags:
        - Events
      summary: Get a single event
      description: Retrieve details of a specific event by its ID.
      operationId: getEventById
      produces:
        - application/json
      parameters:
        - name: event
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: Successful operation
        '404':
          description: Event not found

    post:
      tags:
        - Events
      summary: Update an event
      description: Update an existing event by providing new details.
      operationId: updateEvent
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: event
          in: path
          required: true
          type: integer
        - in: body
          name: event
          description: Updated event details
          required: true
          schema:
            type: object
            properties:
              title:
                type: string
              description:
                type: string
              date:
                type: string
                format: date
      responses:
        '200':
          description: Event updated successfully
        '400':
          description: Bad request
        '404':
          description: Event not found

    delete:
      tags:
        - Events
      summary: Delete an event
      description: Remove an event by its ID.
      operationId: deleteEvent
      produces:
        - application/json
      parameters:
        - name: event
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: Event deleted successfully
        '404':
          description: Event not found
...
securityDefinitions:
  api_key:
    type: apiKey
    name: Authorization
    in: header